Use rust_decimal::Decimal for Money & fin calcs (#143)
* adds rusty-money crate

* log db connection attempt

* move Money type to own file

* use rust_decimal

* helper decimal_at_least_zero

* [user] change numeric fields to Decimal/Money

* money lib

* use decimal for money types in models

* change to decimal in controllesr

* use decimal in finchplaid mod

* use decimal in goals service

* use decimal in snapshots

* use decimal in similar user insights

* use Decimal for `balance` in accountSuccess

* enable serde-float for rust_decimal

+++ server/src/controllers/plaid_controller.rs
@@ -7,6 +7,7 @@ use actix_web::{
   web::{Data, Path},
   HttpResponse,
 };
+use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 use std::sync::{Arc, Mutex};
 
@@ -31,7 +32,7 @@ pub struct AccountResponse {
 pub struct AccountSuccess {
   pub item_id: String,
   pub name: String,
-  pub balance: i64,
+  pub balance: Decimal,
 }
 
 #[derive(Serialize)]
