+++ server/src/models/goal_model.rs
@@ -1,3 +1,4 @@
+use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 use wither::{mongodb::bson::oid::ObjectId, Model};
 
@@ -12,7 +13,7 @@ pub struct Goal {
   #[serde(rename = "end")]
   pub end: i64,
   #[serde(rename = "threshold")]
-  pub threshold: i64,
+  pub threshold: Decimal,
   #[serde(rename = "metric")]
   pub metric: GoalMetrics,
 }
@@ -27,5 +28,5 @@ pub enum GoalMetrics {
 #[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
 pub struct GoalAndStatus {
   pub goal: Goal,
-  pub progress: f64,
+  pub progress: Decimal,
 }
